Add API BatchStartCdnDomain, BatchStopCdnDomain, DescribeTagResources, DescribeUserTags, TagResources, UntagResources.

+++ services/cdn/struct_bps_data_model.go
@@ -20,5 +20,5 @@ type BpsDataModel struct {
 	TimeStamp    string `json:"TimeStamp" xml:"TimeStamp"`
 	LocationName string `json:"LocationName" xml:"LocationName"`
 	IspName      string `json:"IspName" xml:"IspName"`
-	Bps          int    `json:"Bps" xml:"Bps"`
+	Bps          int64  `json:"Bps" xml:"Bps"`
 }
